move more pages to tsx
