--- packages/code-du-travail-frontend-e2e/codecept/features/contributions/contribution.feature
@@ -1,15 +1,0 @@
-#language: fr
-
-@contributions.read
-Fonctionnalité: Contributions
-  Pour consulter une contribution
-  En tant que visiteur
-
-  Scénario:
-    Soit un utilisateur sur la page "/contribution/quelle-peut-etre-la-duree-maximale-dun-cdd"
-
-    Alors je vois "Que dit le code du travail"
-    Alors je vois "Texte applicable"
-    Alors je vois "Références"
-    Alors je vois "L1242-8-1"
-    Alors je vois "Que dit votre convention collective"
